migrate swift4

+++ UITestApplicationUITests/UITestApplicationTableUITests.swift
@@ -24,7 +24,7 @@ class UITestApplicationTableUITests: XCTestCase {
         let rowCount = 6
         
         for i in 0..<rowCount {
-            let cell = cells.element(boundBy: UInt(i))
+            let cell = cells.element(boundBy: i)
             let label = cell.staticTexts[i.description]
             // test will be failed
             XCTAssertTrue(label.exists, "row number:" + i.description)
